Avoid -Werror,-Wdeprecated-anon-enum-enum-conversion
...with Clang 10 trunk:

> vcl/quartz/salgdicommon.cxx:1299:59: error: bitwise operation between different enumeration types ('CGImageAlphaInfo' and '(anonymous enum at /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX10.15.sdk/System/Library/Frameworks/CoreGraphics.framework/Headers/CGImage.h:51:9)') is deprecated [-Werror,-Wdeprecated-anon-enum-enum-conversion]
>                                kCGImageAlphaNoneSkipFirst | kCGBitmapByteOrder32Big );
>                                ~~~~~~~~~~~~~~~~~~~~~~~~~~ ^ ~~~~~~~~~~~~~~~~~~~~~~~

<https://developer.apple.com/documentation/coregraphics/
1455939-cgbitmapcontextcreate> documents the "uint32_t bitmapInfo" parameter of
CGBitmapContextCreate as:  "The constants for specifying the alpha channel
information are declared with the CGImageAlphaInfo type but can be passed to
this parameter safely.  You can also pass the other constants associated with
the CGBitmapInfo type."

Reviewed-on: https://gerrit.libreoffice.org/c/core/+/86175
Tested-by: Jenkins

+++ vcl/quartz/salgdicommon.cxx
@@ -1296,7 +1296,7 @@ Color AquaSalGraphics::getPixel( long nX, long nY )
     CGContextRef xOnePixelContext =
         CGBitmapContextCreate( &aPixel, 1, 1, 8, 32,
                                GetSalData()->mxRGBSpace,
-                               kCGImageAlphaNoneSkipFirst | kCGBitmapByteOrder32Big );
+                               uint32_t(kCGImageAlphaNoneSkipFirst) | uint32_t(kCGBitmapByteOrder32Big) );
 
     // update this graphics layer
     ApplyXorContext();
